// src/lib/types.ts
export interface CongratulationsCard {
  /** Title will contain the name of the person congratulating Saša */
  title: string;
  /** Message will be some heartwarming sentiment describing how Saša impacted that person's life */
  message: string;
  /** An ISO string representing timestamp of the message sent */
  timestamp: string;
}